--- Assets/Resources/Scripts/Common/Goal.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: ccaae83854e7fe4418e483221448cc25
-folderAsset: yes
-timeCreated: 1505888960
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
