+++ Assets/Resources/Scripts/Common/UI/CountDown.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: fa2b7765ae64deb4499e03cafaaa88b0
-timeCreated: 1505889066
+guid: 1ed7ed6662adef946a5bd6e49e354345
+timeCreated: 1517379635
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
